+++ lib/src/components/adaptive_screen/adaptive_screen.describe.dart
@@ -1,7 +1,0 @@
-part of 'adaptive_screen.dart';
-
-String _describeInstance(
-    BuildContext context, ImpaktfullUiAdaptiveScreen instance) {
-  final descriptor = ComponentDescriptor();
-  return descriptor.describe();
-}