+++ lib/src/components/bottom_navigation/bottom_navigation.dart
@@ -1,15 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:impaktfull_ui/src/components/auto_layout/auto_layout.dart';
 import 'package:impaktfull_ui/src/components/bottom_navigation/bottom_navigation_style.dart';
-import 'package:impaktfull_ui/src/util/descriptor/component_descriptor_mixin.dart';
 import 'package:impaktfull_ui/src/widget/override_components/overridable_component_builder.dart';
 
 export 'bottom_navigation_style.dart';
 
-part 'bottom_navigation.describe.dart';
-
-class ImpaktfullUiBottomNavigation extends StatelessWidget
-    with ComponentDescriptorMixin {
+class ImpaktfullUiBottomNavigation extends StatelessWidget {
   final List<Widget> items;
   final bool removeTop;
   final ImpaktfullUiBottomNavigationTheme? theme;
@@ -44,7 +40,4 @@ class ImpaktfullUiBottomNavigation extends StatelessWidget
       ),
     );
   }
-
-  @override
-  String describe(BuildContext context) => _describeInstance(context, this);
 }