+++ lib/src/components/command_menu/command_menu.dart
@@ -1,15 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:impaktfull_ui/src/components/command_menu/command_menu_style.dart';
 import 'package:impaktfull_ui/src/components/command_menu/commander/commander.dart';
-import 'package:impaktfull_ui/src/util/descriptor/component_descriptor_mixin.dart';
 import 'package:impaktfull_ui/src/widget/override_components/overridable_component_builder.dart';
 
 export 'command_menu_style.dart';
 
-part 'command_menu.describe.dart';
-
-class ImpaktfullUiCommandMenu extends StatefulWidget
-    with ComponentDescriptorMixin {
+class ImpaktfullUiCommandMenu extends StatefulWidget {
   final ShortcutActivator? shortcutActivator;
   final Widget child;
   final Widget Function(
@@ -31,9 +27,6 @@ class ImpaktfullUiCommandMenu extends StatefulWidget
   @override
   State<ImpaktfullUiCommandMenu> createState() =>
       _ImpaktfullUiCommandMenuState();
-
-  @override
-  String describe(BuildContext context) => _describeInstance(context, this);
 }
 
 class _ImpaktfullUiCommandMenuState extends State<ImpaktfullUiCommandMenu> {