+++ lib/src/components/floating_action_button/floating_action_button.describe.dart
@@ -1,8 +1,0 @@
-part of 'floating_action_button.dart';
-
-String _describeInstance(
-    BuildContext context, ImpaktfullUiFloatingActionButton instance) {
-  final descriptor = ComponentDescriptor();
-  descriptor.add('theme', instance.theme);
-  return descriptor.describe();
-}