+++ lib/src/components/fluid_padding/fluid_padding.dart
@@ -1,17 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:impaktfull_ui/src/components/fluid_padding/fluid_padding_breakpoint.dart';
 import 'package:impaktfull_ui/src/components/fluid_padding/fluid_padding_style.dart';
-import 'package:impaktfull_ui/src/util/descriptor/component_descriptor_mixin.dart';
 import 'package:impaktfull_ui/src/util/extension/list_extension.dart';
 import 'package:impaktfull_ui/src/widget/override_components/overridable_component_builder.dart';
 
 export 'fluid_padding_style.dart';
 export 'fluid_padding_breakpoint.dart';
 
-part 'fluid_padding.describe.dart';
-
-class ImpaktfullUiFluidPadding extends StatelessWidget
-    with ComponentDescriptorMixin {
+class ImpaktfullUiFluidPadding extends StatelessWidget {
   final List<ImpaktfullUiFluidPaddingBreakPoint>? breakPoints;
   final double topPadding;
   final double bottomPadding;
@@ -53,7 +49,4 @@ class ImpaktfullUiFluidPadding extends StatelessWidget
       },
     );
   }
-
-  @override
-  String describe(BuildContext context) => _describeInstance(context, this);
 }