+++ lib/src/components/interaction_feedback/touch_feedback_sequence/touch_feedback_sequence.describe.dart
@@ -1,7 +1,0 @@
-part of 'touch_feedback_sequence.dart';
-
-String _describeInstance(
-    BuildContext context, ImpaktfullUiTouchFeedbackSequence instance) {
-  final descriptor = ComponentDescriptor();
-  return descriptor.describe();
-}