+++ lib/src/components/introduction/introduction.describe.dart
@@ -1,8 +1,0 @@
-part of 'introduction.dart';
-
-String _describeInstance(
-    BuildContext context, ImpaktfullUiIntroduction instance) {
-  final descriptor = ComponentDescriptor();
-  descriptor.add('theme', instance.theme);
-  return descriptor.describe();
-}