+++ lib/src/components/markdown/markdown.describe.dart
@@ -1,7 +1,0 @@
-part of 'markdown.dart';
-
-String _describeInstance(BuildContext context, ImpaktfullUiMarkdown instance) {
-  final descriptor = ComponentDescriptor();
-  descriptor.add('theme', instance.theme);
-  return descriptor.describe();
-}