+++ lib/src/components/notification_badge/notification_badge.describe.dart
@@ -1,7 +1,0 @@
-part of 'notification_badge.dart';
-
-String _describeInstance(
-    BuildContext context, ImpaktfullUiNotificationBadge instance) {
-  final sb = StringBuffer();
-  return sb.toString();
-}