+++ lib/src/components/password_strength_indicator/password_strength_indicator.dart
@@ -3,16 +3,12 @@ import 'package:impaktfull_ui/src/components/asset/asset_widget.dart';
 import 'package:impaktfull_ui/src/components/auto_layout/auto_layout.dart';
 import 'package:impaktfull_ui/src/components/password_strength_indicator/model/password_strength_indicator_requirement.dart';
 import 'package:impaktfull_ui/src/components/password_strength_indicator/password_strength_indicator_style.dart';
-import 'package:impaktfull_ui/src/util/descriptor/component_descriptor_mixin.dart';
 import 'package:impaktfull_ui/src/widget/override_components/overridable_component_builder.dart';
 
 export 'password_strength_indicator_style.dart';
 export 'model/password_strength_indicator_requirement.dart';
 
-part 'password_strength_indicator.describe.dart';
-
-class ImpaktfullUiPasswordStrengthIndicator extends StatelessWidget
-    with ComponentDescriptorMixin {
+class ImpaktfullUiPasswordStrengthIndicator extends StatelessWidget {
   final int? strengthIndicatorIndex;
   final List<Color>? strengthIndicators;
   final List<PasswordStrengthIndicatorRequirement> requirements;
@@ -99,9 +95,6 @@ class ImpaktfullUiPasswordStrengthIndicator extends StatelessWidget
       },
     );
   }
-
-  @override
-  String describe(BuildContext context) => _describeInstance(context, this);
 
   Color _getColor(
     ImpaktfullUiPasswordStrengthIndicatorTheme componentTheme,