+++ lib/src/components/switch/switch.describe.dart
@@ -1,9 +1,0 @@
-part of 'switch.dart';
-
-String _describeInstance(BuildContext context, ImpaktfullUiSwitch instance) {
-  final descriptor = ComponentDescriptor();
-  descriptor.add('value', instance.value);
-  descriptor.add(
-      'accessibleNavigation', MediaQuery.accessibleNavigationOf(context));
-  return descriptor.describe();
-}