+++ lib/src/components/tab_bar/tab_bar.dart
@@ -1,14 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:impaktfull_ui/src/components/auto_layout/auto_layout.dart';
 import 'package:impaktfull_ui/src/components/tab_bar/tab_bar_style.dart';
-import 'package:impaktfull_ui/src/util/descriptor/component_descriptor_mixin.dart';
 import 'package:impaktfull_ui/src/widget/override_components/overridable_component_builder.dart';
 
 export 'tab_bar_style.dart';
 
-part 'tab_bar.describe.dart';
-
-class ImpaktfullUiTabBar extends StatefulWidget with ComponentDescriptorMixin {
+class ImpaktfullUiTabBar extends StatefulWidget {
   final TabController controller;
   final List<Widget> items;
   final ImpaktfullUiTabBarTheme? theme;
@@ -22,9 +19,6 @@ class ImpaktfullUiTabBar extends StatefulWidget with ComponentDescriptorMixin {
 
   @override
   State<ImpaktfullUiTabBar> createState() => _ImpaktfullUiTabBarState();
-
-  @override
-  String describe(BuildContext context) => _describeInstance(context, this);
 }
 
 class _ImpaktfullUiTabBarState extends State<ImpaktfullUiTabBar> {