+++ lib/src/components/wrap/wrap.describe.dart
@@ -1,6 +1,0 @@
-part of 'wrap.dart';
-
-String _describeInstance(BuildContext context, ImpaktfullUiWrap instance) {
-  final descriptor = ComponentDescriptor();
-  return descriptor.describe();
-}